requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
psycopg[binary]>=3.1.0
psycopg-pool>=3.1.0
Pillow>=10.0.0
httpx>=0.25.0
google-genai>=1.0.0
sentence-transformers>=2.2.2
